Fit nonlinear regression models

Fit nonlinear regression models

Fit nonlinear regression models

Details

NonLinearFitter

An R6 class for automatically fitting non-linear regression models. Includes a library of pre-defined models to simplify selection.

The choice of fitting method depends on the arguments:

If force_optim = TRUE, the function will use optim() regardless of whether weights are supplied.

If weights is supplied, optim() will be used even if force_optim = FALSE.

If neither force_optim nor weights is supplied, the function defaults to nls() for unweighted fitting.

Behavior Examples:

Default behavior: nls() is used when weights = NULL and force_optim = FALSE.

Weighted fitting: optim() is used when weights is provided, even if force_optim = FALSE.

Forced optimization: optim() is used when force_optim = TRUE, regardless of whether weights is supplied.

Public fields

data

A data.table containing the dataset for modeling.

models

A list of non-linear models to test.

fit_results

A list to store the results of model fits.

evaluation_metrics

A list to store evaluation metrics for each model.

plots

A list to store plots of model fits.

model_library

A pre-defined library of common non-linear models. Initialize the NonLinearFitter class

Methods

Public methods

NonLinearFitter$new()

NonLinearFitter$list_models()

NonLinearFitter$add_model()

NonLinearFitter$fit_models()

NonLinearFitter$model_comparison_plot()

NonLinearFitter$clone()

Method new()

Usage

NonLinearFitter$new(data)

Arguments

data

A data.table containing the dataset for modeling. Must include the predictor and response variable columns.

Returns

A new instance of the NonLinearFitter class.

Method list_models()

Usage

NonLinearFitter$list_models()

Returns

A data.table summarizing available models. Add a non-linear model for testing

Method add_model()

Usage

NonLinearFitter$add_model(name, formula = NULL, start_params = NULL)

Arguments

name

The name of the model (e.g., "Hill").

formula

The non-linear formula for the model (optional if using pre-defined model).

start_params

A list of starting parameters for the model (optional if using pre-defined model).

Returns

NULL

Method fit_models()

fit_models() first standardizes your data before fitting the model. The fitting method depends on whether a weights_col is provided: nls() is used for unweighted fitting, while optim() is used for weighted fitting. The returned parameters are based on the standardized data. However, when scoring models, the results are back-transformed to align with the original data scale.

Usage

NonLinearFitter$fit_models(
  x_col,
  y_col,
  weights_col = NULL,
  control = list(maxiter = 1024),
  force_optim = FALSE,
  ...
)

Arguments

x_col

The name of the predictor variable.

y_col

The name of the response variable.

weights_col

The name of the weights variable.

control

A list of control parameters for the optimizer, such as maxiter. Default is list(maxiter = 200).

force_optim

Logical; if TRUE, forces the use of optim() regardless of whether weights are supplied. Defaults to FALSE.

...

Additional arguments to be passed to the underlying fitting functions (nlsLM for unweighted models or optim for weighted models). Examples include trace, lower, and upper for nlsLM, or reltol, parscale, and others for optim.

Returns

A list of fitted model objects. Generate a comparison plot for model shapes

Method model_comparison_plot()

Usage

NonLinearFitter$model_comparison_plot(
  x_range = seq(1, 100, by = 1),
  normalize = TRUE,
  theme = "macarons"
)

Arguments

x_range

A numeric vector specifying the range of x values to evaluate.

normalize

Logical. If TRUE, normalizes the y values for each model to fall between 0 and 1. Defaults to TRUE.

theme

A string specifying the plot theme (e.g., "macarons").

Returns

An echarts4r object representing the comparison plot.

Method clone()

The objects of this class are cloneable with this method.

Usage

NonLinearFitter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.