NonLinearModelEvaluator

An R6 class to evaluate non-linear regression models. Includes tools to generate tables of statistics and visualizations to compare models against data.

Public fields

fit_results

A list of fitted model objects.

evaluation_metrics

A data.table containing model evaluation metrics.

plots

A list of visualizations comparing models against data.

data

The original dataset used for fitting models. Initialize the NonLinearModelEvaluator class

Methods

Public methods

NonLinearModelEvaluator$new()

NonLinearModelEvaluator$generate_metrics()

NonLinearModelEvaluator$generate_comparison_plot()

NonLinearModelEvaluator$clone()

Method new()

Usage

NonLinearModelEvaluator$new(fit_results, data)

Arguments

fit_results

A list of fitted model objects (e.g., output from NonLinearFitter).

data

The original dataset used for fitting models.

Returns

A new instance of the NonLinearModelEvaluator class.

Method generate_metrics()

Usage

NonLinearModelEvaluator$generate_metrics(y_col = NULL, x_col = NULL)

Arguments

y_col

target variable

x_col

x variable

Returns

A data.table of evaluation metrics with fitted equations.

Method generate_comparison_plot()

Usage

NonLinearModelEvaluator$generate_comparison_plot(
  data,
  x_col,
  y_col,
  theme = "macarons"
)

Arguments

data

A data.table or data.frame containing the dataset used for evaluation.

x_col

A string specifying the name of the x variable in the dataset.

y_col

A string specifying the name of the y variable in the dataset.

theme

Echarts theme

Returns

An echarts4r plot showing observed vs. predicted data, with weighted comparisons if available.

Method clone()

The objects of this class are cloneable with this method.

Usage

NonLinearModelEvaluator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.