NonLinearModelScorer

An R6 class to score non-linear regression models on new data and visualize the results.

Public fields

fit_results

A list of fitted model objects.

scored_data

A list of data.tables containing scored data.

score_plots

A list of plots visualizing scored data.

Methods

Public methods

NonLinearModelScorer$new()

NonLinearModelScorer$score_new_data()

NonLinearModelScorer$generate_score_plot()

NonLinearModelScorer$clone()

Method new()

Usage

NonLinearModelScorer$new(fit_results)

Arguments

fit_results

A list of fitted model objects (e.g., output from NonLinearFitter).

Returns

A new instance of the NonLinearModelScorer class.

Method score_new_data()

Usage

NonLinearModelScorer$score_new_data(new_data, x_col)

Arguments

new_data

A data.table containing the new data to score.

x_col

The predictor column in new_data.

Returns

A list of data.tables with predicted values for each model.

Method generate_score_plot()

Usage

NonLinearModelScorer$generate_score_plot(
  model_name,
  new_data,
  x_col,
  theme = "westeros"
)

Arguments

model_name

The name of the model to plot.

new_data

The original new data used for scoring.

x_col

The predictor column in new_data.

theme

Echarts theme

Returns

A plot visualizing the scored data.

Method clone()

The objects of this class are cloneable with this method.

Usage

NonLinearModelScorer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.